fix: 잼만들기 모달 로직 페이지로 이동
